icn_api::dag_trait

Trait DagApi

pub trait DagApi {
    // Required method
    fn get_dag_root<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Cid>, CommonError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

API surface for DAG operations.

Required Methods

fn get_dag_root<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Option<Cid>, CommonError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the current root CID of the node’s DAG.

Clients can poll this value to detect when a peer has new history and needs state synchronization.